requirements.txt
langchain>=0.0.325
langgraph>=0.0.15
python-docx>=0.8.11
requests>=2.31.0
openai>=1.3.5
